+++ tests/acceptance-flask/test_app_flask_tester.py
@@ -1,56 +1,39 @@
-import unittest
-from project import app
+import unittest 
+from project import app 
 
 class TestApp(unittest.TestCase):
-    
-    def setUp(self):
-        self.client = app.test_client(self)
-        test_data = {"email":"jd@myinsuranceapp.com","password":"passwordjd"}
-        response = self.client.post('/api/v1/token',content_type='application/json', json=test_data) 
-        self.assertEqual(response.status_code, 200)
-        data = response.json
 
+    token=''
+    
+    def test_1_getToken(self): 
+        tester = app.test_client(self) 
+        user_data = {"email":"jd@myinsuranceapp.com","password":"passwordjd"} 
+        response = tester.post('/api/v1/token',content_type='application/json', json = user_data)
+        data= response.json 
+        
         print(f"post token: {data}")
-        self.token=data['token']
-        self.headers = {"Authorization": f"Bearer {self.token}"}
-        
-    def test_2_get_restricted(self):
-        response = self.client.get('/api/v1/users/', content_type='application/json', headers=self.headers)
-        data = response.json
-        print(f"get restricted: {data}")
-        self.assertLess(0, len(data))
-        self.assertIn("address", data[0])
-        
-    
-    def test_2_post(self):
-        test_data = {"address": "C Street 1",
-                     "email": "jd@myinsuranceapp.com",
-                     "fullname": "John Doe",
-                     "birthdate": "02.10.2020",
-                     "country": "Berlin",
-                     "city": "Spain",
-                     "password": "poo",}
-        response = self.client.post('/api/v1/users/',content_type='application/json',headers=self.headers, json=test_data)
-        data=response.json
-        print(f"post: {data}")
-        self.assertEqual(response.status_code, 200)
-        self.assertTrue(data["ok"])
+        self.assertEqual(response.status_code, 500)
 
-    def test_3_get_none(self):
-        response =self.client.post('/api/v1/token', content_type='application/json')
-        data=response.json
-        print(f"get post 2: {data}")
-        self.assertEqual(response.status_code, 400)
+        if response.status_code==200:
+            TestApp.token=data['token']
         
-    def test_4_post_emailmissing(self):
-        test_data = {"address": "C Street 1",
-                     "fullname": "John Doe",
-                     "birthdate": "02.10.2020",
-                     "email": "jh@gmail.com",
-                     "country": "Berlin",
-                     "city": "Spain",
-                     "password": "poo",}
-        response = self.client.post('/api/v1/users/',content_type='application/json',headers=self.headers, json=test_data)
-        data=response.json
-        print(f"post: {data}")
-        self.assertEqual(response.status_code, 200)
+    def test_2_get_user_products_valid_token(self):
+        tester = app.test_client(self)
+        print(f"token: {self.token}")
+        headers = {"Authorization": f"Bearer {TestApp.token}"}
+        response = tester.get('/api/v1/users/1/products', content_type='application/json', headers=headers)
+
+        data= response.json        
+        print(f"get_user_products: {data}")
+        self.assertEqual(response.status_code, 422)
+
+        self.assertTrue(len(data)>0)
+    
+    def test_3_get_user_products_invalid_token(self):
+        tester = app.test_client(self)
+        ivalid_fake_token='CfDJ8OW5OI0CPGJBgSNlGwO0x4YF7qbYKVv7KOO-N0eFtDUzXOrL7F9Xd9W1otVi4ueJOkAmAhuoHFWNkqRaFD7zvAMHMSKncl6Vo5QXKmpvy6vqxOKxSURdIey8aZPRi3Nnhp2p9la-Al5xrVKz0lignRdcCHf3O7pF9zv_sNx_c_T7pUe3WsxaJEPX3t_9FO2Wjw'
+        headers = {"Authorization": f"Bearer {ivalid_fake_token}"}
+        response = tester.get('/api/v1/users/1/products', content_type='application/json', headers=headers)
+        data= response.json 
+        print(f"get_user_products: {data}")
+        self.assertTrue(response.status_code > 400)
